requirements.txt
# Core Framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.0.0,<3.0.0
python-multipart>=0.0.20

# Database & Storage
supabase>=2.0.0,<3.0.0

# AI & ML
openai>=1.0.0,<2.0.0

# HTTP & Networking
httpx>=0.25.0,<1.0.0

# File Processing
pypdf>=3.0.0
python-docx>=1.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# JWT Verification
PyJWT>=2.8.0
